Bugfix for cumulative elevation gain
Fixed a bug with stats calculation. If the first point in the segment
had an elevation of null, that would cause the elevation gain to
increase from NULL (or zero) to the next point's elevation. Added a
check to see, whether there is elevation data in the point and only then
add it.